pyproject.toml
[tool.poetry]
name = "toy"
version = "0.0.1"
description = ""
authors = ["R202"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
flask-mongoengine = "^1.0.0"
uWSGI = "^2.0.19"
Werkzeug = "^2.0.1"
Flask-Login = "^0.5.0"
Flask-Principal = "^0.4.0"
watchdog = "^2.1.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.7b0"

[tool.poetry.scripts]
toy-web = "toy.cmd.web:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
